% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{apply_styling}
\alias{apply_styling}
\title{Apply Spreadsheet Formatting to gt Table}
\usage{
apply_styling(gt_object, formatRdy)
}
\arguments{
\item{gt_object}{A gt object.}

\item{formatRdy}{A data frame with formatting definitions.}
}
\value{
A great table!
}
\description{
Applies formatting to a gt table based on information extracted from the spreadsheet file.
}
\examples{
example_spreadsheet <- system.file("extdata/rodentsheet.xlsx", package = "forgts")
sprdsht_data <- readxl::read_excel(example_spreadsheet)
gt_table <- gt::gt(sprdsht_data)
format_long <- get_formatting(example_spreadsheet)
format_ready <- translate_defs(format_long)
apply_styling(gt_table, format_ready)

}
