% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forgts-package.R
\docType{package}
\name{forgts-package}
\alias{forgts-package}
\title{forgts: Convert Formatted Spreadsheets to Presentation-Ready Display Tables}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Reads data plus formatting from a spreadsheet file and creates a gt object with the same data and formatting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/luisDVA/forgts}
  \item \url{https://luisdva.github.io/forgts/}
  \item Report bugs at \url{https://github.com/luisDVA/forgts/issues}
}

}
\author{
\strong{Maintainer}: Luis D. Verde Arregoitia \email{luis@liomys.mx} (\href{https://orcid.org/0000-0001-9520-6543}{ORCID}) [copyright holder]

}
\keyword{internal}
