% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forgts_main.R
\name{forgts}
\alias{forgts}
\title{Creat gt from  spreadsheet}
\usage{
forgts(file, sheet = NULL)
}
\arguments{
\item{file}{Path to a spreadsheet.}

\item{sheet}{A character string with the name of a sheet within a workbook,
or an integer with the position of the sheet. Defaults to the first sheet
if left unspecified.}
}
\value{
A great table!
}
\description{
Wrapper. Reads a spreadsheet file including most formatting information and creates a
formatted gt table with the same data and formatting.
}
\examples{
\dontrun{
forgts("yourpath/yourspreadsheet.xlsx")
}
}
