% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formatting.R
\name{get_formatting}
\alias{get_formatting}
\title{Get Spreadsheet Formatting}
\usage{
get_formatting(xlfilepath, sheet = NULL)
}
\arguments{
\item{xlfilepath}{A character string specifying the path to the spreadsheet.}

\item{sheet}{A character string with the name of a sheet within a workbook,
or an integer with the position of the sheet. Defaults to the first sheet
if left unspecified.}
}
\value{
A data frame or list with parsed formatting information.
}
\description{
Extracts formatting information from a spreadsheet.
}
\examples{
example_file <- system.file("extdata/rodentsheet.xlsx", package = "forgts")
get_formatting(example_file)

}
