% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\name{rodentsheet.xlsx}
\alias{rodentsheet.xlsx}
\title{rodentsheet.xlsx spreadsheet}
\description{
Open XML Format Spreadsheet with 1 sheet, 5 columns, and 8 rows. Toy dataset
with data on measurements for rodent specimens, extracted from Verde
Arregoitia et al. (2017; https://doi.org/10.5281/zenodo.201147). Cells and
cell content have formatting added at random for demonstration purposes.
}
