% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formatting.R
\name{target_var_fmt}
\alias{target_var_fmt}
\title{Colwise formatting}
\usage{
target_var_fmt(format_joined, spsheet, col_name)
}
\arguments{
\item{format_joined}{A data frame produced internally by \code{get_formatting}}

\item{spsheet}{Data frame version of spraedsheet.}

\item{col_name}{Variable to operate on.}
}
\value{
Modified data frame in long format.
}
\description{
Colwise formatting
}
