% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_formatting.R
\name{translate_defs}
\alias{translate_defs}
\title{Translate Formatting Definitions}
\usage{
translate_defs(format_long)
}
\arguments{
\item{format_long}{Data frame produced by \code{get_formatting} containing raw formatting information.}
}
\value{
A data frame with formatting information compatible with gt.
}
\description{
Convert formatting information into arguments that can be used by \code{gt}.
}
\examples{
\dontrun{
format_long <- get_formatting("yourpath/yourspreadsheet.xlsx")
translate_defs(format_long)
}
}
