% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{apply_styling}
\alias{apply_styling}
\title{Apply Spreadsheet Formatting to gt Table}
\usage{
apply_styling(gt_object, formatRdy)
}
\arguments{
\item{formatRdy}{A data frame with formatting definitions.}

\item{gt_table}{A gt object.}
}
\value{
A great table!
}
\description{
Applies formatting to a gt table based on information extracted from the spreadsheet file.
}
\examples{
\dontrun{
sprdsht_data <- read_excel("yourpath/yourspreadsheet.xlsx")
gt_table <- gt::gt(sprdsht_data)
format_long <- get_formatting("yourpath/yourspreadsheet.xlsx")
format_ready <- translate_defs(format_long)
apply_styling(gt_table, format_ready)
}

}
